refactor(mcp): split the use tool's runner half into use-runner.ts
use.ts had grown to 543 lines carrying BOTH the asker surface and the whole runner half
(serve/answer/lease/dispatch) since the Maker fold. Extract the runner half into
mcp-tools/use-runner.ts as runnerDispatch(tc, args) — it returns a tool result for a
REPORT/PULL op, or null to fall through to the asker path. use.ts is now the give/ask
surface plus a thin `if (registered) runnerDispatch(...)` call (543 -> 396 lines); the
runner half is ~190. The shared transcript clipper (clipSessionText/ANSWER_MAX/ENTRY_MAX)
moves to mcp-util.

Matches the OSS "hoist the work out of the handler" norm (github-mcp-server, the MCP
reference servers). Pure refactor, no behavior change: full gate green (ci + typecheck +
1143 api tests + E2E maker loop).
