+++ apps/api/src/mcp-util.ts
@@ -75,6 +75,16 @@ export const clipDoc = (s: string, sections: OutlineSection[]) => {
   return `${s.slice(0, MAX_CHARS)}\n\n…[truncated ${s.length - MAX_CHARS} of ${s.length} chars — ${steer}]`
 }
 
+// The context-session transcript clipper (used by `use`'s asker reply + runner serve):
+// a generous cap on a settled answer, a tight one per transcript entry — together under
+// clipDoc's MAX_CHARS ceiling — with a steer to the console, which holds the full transcript.
+export const ANSWER_MAX = 40_000
+export const ENTRY_MAX = 1_500
+export const clipSessionText = (s: string, max: number, consoleUrl: string): string =>
+  s.length > max
+    ? `${s.slice(0, max)}\n\n…[truncated ${s.length - max} of ${s.length} chars — full transcript in the console: ${consoleUrl}]`
+    : s
+
 // A content-bearing response: a frontmatter-style header, a blank line, then the RAW
 // body — one text block, real newlines, never JSON-escaped. When a client spills it
 // to a file, that file is line-oriented and greppable (the old JSON envelope turned a
